requirements-dev.txt
-r requirements.txt
pytest-django
flake8